--- data/test_db_connection.py
@@ -1,6 +1,0 @@
-import os
-PRIVATE_KEY = os.getenv('PRIVATE_KEY')
-PUBLIC_KEY = os.getenv('PUBLIC_KEY')
-
-print(PRIVATE_KEY)
-print(PUBLIC_KEY)
