+++ routers/api_user.py
@@ -1,15 +1,17 @@
 from fastapi import APIRouter, Request, Form
 from fastapi.responses import JSONResponse
 import mysql.connector
-from datamodel import DataModel
+from pydantic import ValidationError
+from utils.datamodel import DataModel
+
 
 router = APIRouter()
 headers = {"Content-Type": "application/json; charset=utf-8"}
 
 @router.post("/api/user") # data: DataModel = Depends() 這方式得不到資料
 async def api_user_signup(request: Request, name: str = Form(...), email: str = Form(...), password: str = Form(...)):
-    data_formate = DataModel(name=name, email=email, password=password)
     try:
+        data_formate = DataModel(name=name, email=email, password=password)
         db_pool = request.state.db_pool.get("basic_db") 
         with db_pool.get_connection() as connection:
             with connection.cursor() as cursor:
@@ -25,9 +27,21 @@ async def api_user_signup(request: Request, name: str = Form(...), email: str = 
                 input_data = {"error": True,"message": "Invalid registration, duplicate email or other reasons"}
                 return JSONResponse(status_code=400,content=input_data, headers=headers)
             
-    except (mysql.connector.Error, ValueError, Exception) as err:
+    except (mysql.connector.Error) as err:
         return JSONResponse(
             status_code=500,
+            content={"error": True, "message": str(err)},
+            headers=headers
+        )
+    except ValidationError as err:
+        return JSONResponse(
+            status_code=422,
+            content={"error": True, "message": err.errors()}, # .errors()可以返回更仔細的資料
+            headers=headers
+        )      
+    except (ValueError, Exception) as err:
+        return JSONResponse(
+            status_code=400,
             content={"error": True, "message": str(err)},
             headers=headers
         )
