+++ static/attraction.html
@@ -3,15 +3,25 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Attraction</title>
     <link
       rel="stylesheet"
       href="/static/css_js_folder/css/attraction_style.css"
     />
+    <link
+      rel="stylesheet"
+      href="/static/css_js_folder/css/signin_signout_section.css"
+    />
 
-    <title>Attraction</title>
+    <script src="/static/css_js_folder/js/attraction_script.js" defer></script>
+    <script
+      src="/static/css_js_folder/js/signin_signout_section.js"
+      defer
+    ></script>
   </head>
   <body>
     <div class="content">
+      <div id="overlay" hidden onclick="cancelFormDisplay()"></div>
       <!-- ! navbar part start -->
       <div class="nav-bar">
         <div class="second-nav-bar">
@@ -20,11 +30,118 @@
             <div class="firstbtn">
               <a href="#" class="first-bar-btn">預定行程</a>
             </div>
-            <div class="secondbtn">
-              <a href="#" class="second-bar-btn">登入/註冊</a>
+            <div class="secondbtn" id="secondbtn">
+              <a
+                class="second-bar-btn"
+                id="switch-btn"
+                onclick=" displaySignIn()"
+                >登入/註冊</a
+              >
             </div>
           </div>
         </div>
+      </div>
+      <!-- ! signin-form -->
+      <div hidden class="signin-form-div" id="signin-form-div">
+        <p class="form-text">登入會員帳號</p>
+        <button
+          class="log-img-button"
+          id="log-in-img-button"
+          onclick="cancelFormDisplay()"
+        >
+          <img
+            src="https://raw.githubusercontent.com/RaphaFang/taipei-day-trip/develop/static/css_js_folder/pic/X.png"
+            alt="Image Button"
+          />
+        </button>
+        <form
+          id="signin-form"
+          action="javascript:void(0);"
+          onsubmit="submitSigninForm()"
+        >
+          <div class="form-group">
+            <input
+              type="email"
+              id="signin_email"
+              name="email"
+              placeholder="輸入電子信箱"
+              class="for-input-padding"
+            />
+          </div>
+          <div class="form-group">
+            <input
+              type="password"
+              id="signin_password"
+              name="password"
+              placeholder="輸入密碼"
+              class="for-input-padding"
+            />
+          </div>
+          <div class="form-group-last">
+            <button type="submit">登入帳戶</button>
+          </div>
+          <div class="switch-form" id="switch-form">
+            <a onclick="signUpInSwitch()">還沒有帳戶？點此註冊</a>
+          </div>
+          <div hidden class="error-catcher" id="error-catcher"></div>
+        </form>
+      </div>
+      <!-- ! signup form -->
+      <div hidden class="signup-form-div" id="signup-form-div">
+        <p class="form-text">註冊會員帳號</p>
+        <button
+          class="log-img-button"
+          id="log-in-img-button"
+          onclick="cancelFormDisplay()"
+        >
+          <img
+            src="https://raw.githubusercontent.com/RaphaFang/taipei-day-trip/develop/static/css_js_folder/pic/X.png"
+            alt="Image Button"
+          />
+        </button>
+        <form
+          id="signup-form"
+          action="javascript:void(0);"
+          onsubmit="submitSignUpForm()"
+        >
+          <div class="form-group">
+            <input
+              type="text"
+              id="signup_username"
+              name="name"
+              placeholder="輸入姓名"
+              required
+              class="for-input-padding"
+            />
+          </div>
+          <div class="form-group">
+            <input
+              type="email"
+              id="signup_user_email"
+              name="email"
+              placeholder="輸入電子郵件"
+              required
+              class="for-input-padding"
+            />
+          </div>
+          <div class="form-group">
+            <input
+              type="password"
+              id="signup_password"
+              name="password"
+              placeholder="輸入密碼"
+              required
+              class="for-input-padding"
+            />
+          </div>
+          <div class="form-group-last">
+            <button type="submit" class="form-btn">註冊新帳戶</button>
+          </div>
+          <div class="switch-form" id="switch-form">
+            <a onclick="signUpInSwitch()">已經有帳戶了？點此登入</a>
+          </div>
+          <div hidden class="error-catcher" id="error-catcher"></div>
+        </form>
       </div>
       <div class="without-nav-bar">
         <!-- ! pic and day select -->
@@ -118,6 +235,5 @@
     <footer class="foot">
       <p class="foot-p">COPYRIGHT © 2021 台北一日遊</p>
     </footer>
-    <script src="/static/css_js_folder/js/attraction_script.js"></script>
   </body>
 </html>
