+++ static/css_js_folder/css/index_style.css
@@ -170,7 +170,6 @@ body {
   max-width: 400px;
   padding-left: 10px;
 
-  filter: drop-shadow(0px 0px 20px #aabbcc);
   background: #ffffff;
 }
 
@@ -484,24 +483,13 @@ body {
   text-align: left;
   color: rgba(255, 255, 255, 1);
 }
-
 /* ! RWD */
-@media (max-width: 1200px) {
-}
-
-@media (max-width: 900px) {
-}
-
-@media (max-width: 600px) {
-}
-
 @media (max-width: 360px) {
   .second-nav-bar {
     width: 100%;
     max-width: 360px;
     padding-left: 10px;
   }
-
   .title {
     padding-left: 10px;
     font-family: Noto Sans TC;
@@ -511,15 +499,12 @@ body {
     text-align: left;
     white-space: nowrap;
   }
-
   .hero-background-image-container {
-    /* width: 100%; */
     width: 360px;
     height: 320px;
     top: -217px;
     left: -1469px;
     gap: 0px;
-    /* opacity: 0; */
   }
   .searching-block {
     width: 340px;
@@ -566,17 +551,12 @@ body {
     height: 100%;
     object-fit: cover;
   }
-
   .attraction-name-div {
     width: 322px;
     height: 40px;
     top: 195px;
   }
 }
-
 .background-image-container {
   pointer-events: auto;
-}
-
-.form-div {
 }
