+++ static/css_js_folder/css/signin_signout_section.css
@@ -1,0 +1,132 @@
+#overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 1);
+  z-index: 1000;
+  opacity: 25%;
+}
+.signup-form-div,
+.signin-form-div {
+  width: 340px;
+  min-height: 275px;
+  height: auto;
+  position: fixed;
+  top: 80px; /* 距離頂部 80px */
+  left: 50%;
+  transform: translateX(-50%); /* 只在水平（X 軸）方向上平移 */
+  z-index: 1001; /* 確保在灰色背景之上 */
+  background-color: rgba(255, 255, 255, 1); /* 白色背景 */
+  padding: 25px 15px 15px 15px; /* 添加內邊距讓內容不貼邊 */
+  border-radius: 10px; /* 可選：添加圓角 */
+  box-shadow: 0px 4px 60px 0px rgba(170, 170, 170, 1); /* 添加陰影 */
+}
+.signup-form-div {
+  min-height: 332px;
+  height: auto;
+}
+.signup-form-div:before,
+.signin-form-div:before {
+  content: "";
+  display: block;
+  width: 100%;
+  height: 10px; /* 設置頂邊框的高度 */
+  background: linear-gradient(270deg, #337788 0%, #66aabb 100%);
+  position: absolute;
+  top: 0;
+  left: 0;
+  border-radius: 10px 10px 0 0;
+}
+.log-img-button {
+  position: absolute;
+  top: 17px;
+  right: 17px; /* 距離右側 10px */
+  background: none; /* 移除按鈕背景 */
+  border: none; /* 移除按鈕邊框 */
+  cursor: pointer; /* 鼠標樣式 */
+}
+.log-img-button img {
+  width: 16px;
+  height: 16px;
+}
+.form-text {
+  font-family: Noto Sans TC;
+  font-size: 24px;
+  font-weight: 700;
+  line-height: 24px;
+  text-align: center;
+  width: 310px;
+  height: 27px;
+  margin-top: 0px;
+  color: rgba(102, 102, 102, 1);
+}
+.form-group {
+  width: 310px;
+  height: 47px;
+  margin-top: 10px; /* 使用 auto 使其水平置中，頂部邊距 15px */
+}
+.form-group:first-of-type {
+  margin-top: 15px; /* 設定第一個 form-group 的 margin-top 為 15px */
+}
+.form-group input {
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box; /* 確保寬高包括內邊距 */
+  border: 1px solid rgba(204, 204, 204, 1);
+  border-radius: 5px; /* 可選：添加圓角 */
+  padding: 15px;
+}
+
+.form-group input::placeholder {
+  color: rgba(117, 117, 117, 1);
+  font-family: Noto Sans TC;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 13.3px;
+  text-align: left;
+}
+.form-group input:focus {
+  outline: none;
+  border: 1px solid rgba(204, 204, 204, 1);
+}
+
+.form-group-last button {
+  width: 100%;
+  height: 47px;
+  margin-top: 10px;
+
+  border-radius: 5px; /* 可選：添加圓角 */
+  background-color: rgba(68, 136, 153, 1);
+  border: 0;
+  color: white; /* 設置按鈕文字顏色 */
+
+  cursor: pointer; /* 設置鼠標樣式 */
+
+  font-family: Noto Sans TC;
+  font-size: 19px;
+  font-weight: 400;
+  line-height: 16px;
+  text-align: center;
+}
+.error-catcher,
+.switch-form {
+  height: 22px;
+  margin-top: 15px;
+  font-family: Noto Sans TC;
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 13.3px;
+  text-align: center;
+  color: rgba(102, 102, 102, 1);
+}
+#switch-form a {
+  font-size: 16px;
+  color: rgba(102, 102, 102, 1);
+  text-decoration: none;
+}
+
+.error-catcher {
+  margin-bottom: 10px;
+}
